requirements.doc.txt
Sphinx
sphinx-bootstrap-theme
sphinxcontrib-napoleon
GitPython
semantic_version
